Debugging try 1

+++ util/conf_util.py
@@ -1,12 +1,12 @@
 import yaml
 import numpy as np
-from math import comb
+from scipy.special import comb
 
 def get_dimen(n,r,m):
 	code_dimension = 0
 	for w in range(0,r+1):
 		code_dimension += comb(n,w)*np.power((n-1),w)
-	return code_dimension
+	return np.int64(code_dimension)
 
 def get_default_conf(conf_path=None):
 	if conf_path is None:
@@ -24,7 +24,7 @@ def get_default_conf(conf_path=None):
 	conf["test"]["logger_name"] = conf["test"]["logger_name"].format(data_type)
 	
 
-	conf["para"]["code_length"] = np.power(n,m)
+	conf["para"]["code_length"] = np.int64(np.power(n,m))
 	conf["para"]["code_dimension"] = get_dimen(n,r,m)
 
 	return conf
